HERO V VILLAIN

+++ U5_CLASS_WRITING/HeroVillainExample/Power.java
@@ -3,16 +3,23 @@ package U5_CLASS_WRITING.HeroVillainExample;
 public class Power {
     private String power;
 
-    public int powerStrength;
+    public double powerStrength;
 
     public String toString(){
         String output = "";
+
 
         output += "Name of the superpower: " + this.power + "\n";
         output += "Power Strength: " + this.powerStrength;
         return output;
     }
-    public Power(String power, int powerStrength) {
+
+    /**
+     * full constructor of power object
+     * @param power name of superpower
+     * @param powerStrength this is the strength of your power (1-10)
+     */
+    public Power(String power, double powerStrength) {
         this.power = power;
         this.powerStrength = powerStrength;
     }
@@ -25,11 +32,11 @@ public class Power {
         this.power = power;
     }
 
-    public int getPowerStrength() {
+    public double getPowerStrength() {
         return powerStrength;
     }
 
-    public void setPowerStrength(int powerStrength) {
+    public void setPowerStrength(double powerStrength) {
         this.powerStrength = powerStrength;
     }
 }
